axiom_hive/src/axiom_hive/formal_proof.py
"""
Hybrid Proof Verifier: SymPy + Z3 for LHC.
"""

from sympy import symbols, And, Abs, simplify, satisfiable as sympy_sat
from z3 import *
from typing import Dict, Any

class ProofVerifier:
    """
    Verifies LHC symbolically with SymPy/Z3 hybrid.
    """
    def __init__(self):
        self.sympy_proofs: Dict[str, Any] = {}
        self.z3_proofs: Dict[str, Any] = {}

    def load_proofs(self, yaml_config: Dict[str, str]):
        x = symbols('x')
        for name, expr_str in yaml_config['sympy_proofs'].items():
            expr = eval(expr_str)  # Prod: ast.safe
            self.sympy_proofs[name] = simplify(expr)
            if not sympy_sat(self.sympy_proofs[name]):
                raise ValueError(f"SymPy Unsatisfiable: {name}")

        s = Solver()
        x_z3 = Real('x')
        for name, expr_str in yaml_config['z3_proofs'].items():
            expr = eval(expr_str.replace('x', 'x_z3'))  # Prod: safe parse
            s.add(expr)
            if s.check() == unsat:
                raise ValueError(f"Z3 Unsatisfiable: {name}")
            self.z3_proofs[name] = expr

    def verify_sympy(self, name: str) -> bool:
        return bool(sympy_sat(self.sympy_proofs[name]))

    def verify_z3(self, name: str) -> bool:
        s = Solver()
        s.add(Not(self.z3_proofs[name]))
        return s.check() == unsat  # Proof by contradiction

    def verify(self, name: str) -> bool:
        return self.verify_sympy(name) and self.verify_z3(name)