axiom_hive/src/axiom_hive/sanctity.py
"""
Sanctity Protocol: 2025 LHC with SymPy/Z3 hybrid proofs.
"""

import yaml
import torch
from typing import Dict, Any, List, Callable
from .dag import AxiomDAG
from .formal_proof import ProofVerifier

class LexHumanaCorpus:
    """
    LHC: 2025 fused as lambdas + hybrid proofs.
    """
    def __init__(self, yaml_path: str = "lhc.yaml"):
        with open(yaml_path, 'r') as f:
            config = yaml.safe_load(f)
        self.constraints: Dict[str, Callable] = {}
        for name, func_str in config['constraints'].items():
            self.constraints[name] = eval(func_str)  # Prod: safe
        self.verifier = ProofVerifier()
        self.verifier.load_proofs(config)

    def validate(self, outputs: List[torch.Tensor]) -> bool:
        # Empirical
        for constraint in self.constraints.values():
            if not all(constraint(out) for out in outputs):
                return False
        # Hybrid Symbolic
        for name in self.verifier.sympy_proofs:
            if not self.verifier.verify(name):
                return False
        return True

class GovernorCompiler:
    """
    Audits empirically + hybrid symbolically.
    """
    def __init__(self, lhc: LexHumanaCorpus):
        self.lhc = lhc

    def audit_paths(self, dag: AxiomDAG) -> bool:
        try:
            results = dag.execute()
            outputs = [results[node] for node in results if isinstance(results[node], torch.Tensor)]
            return self.lhc.validate(outputs)
        except Exception:
            return False

class SanctityProtocol:
    """
    Wraps with 2025 verifiability.
    """
    def __init__(self, lhc_path: str = "lhc.yaml"):
        self.lhc = LexHumanaCorpus(lhc_path)
        self.governor = GovernorCompiler(self.lhc)

    def sanctify(self, dag: AxiomDAG) -> Dict[str, Any]:
        if self.governor.audit_paths(dag):
            blueprint = {"status": "SANCTIFIED", "blueprint": dag.execute(), "proofs": {k: "Verified" for k in self.lhc.verifier.sympy_proofs}}
            return blueprint
        else:
            raise ValueError("NULL_PATH: Action violates 2025 LHC constraints.")

def execute_sanctified(dag: AxiomDAG, protocol: SanctityProtocol) -> Dict[str, Any]:
    return protocol.sanctify(dag)